requirements.txt
# Function dependencies, for example:
# package>=version
httplib2==0.20.4
google-cloud-logging